added CRUD fior rows in postgres and mysql
