+++ modules/data-viewer/TableViewer.tsx
@@ -13,6 +13,7 @@ import {
   Check,
   XCircle,
   Database,
+  TableProperties,
 } from "lucide-react";
 import React, { useState, useEffect, useCallback, useMemo } from "react";
 import { cn } from "@/lib/utils";
@@ -92,129 +93,9 @@ import {
   deleteRow,
 } from "@/app/actions/data";
 import dynamic from "next/dynamic";
+import { AddRowDialog } from "./AddRowDialog";
 
 const JsonViewer = dynamic(() => import("./JsonViewer"), { ssr: false });
-
-// It's recommended to move AddRowDialog to its own file (e.g., /modules/data-viewer/AddRowDialog.tsx)
-const AddRowDialog = ({
-  isOpen,
-  onClose,
-  schema,
-  connection,
-  tableName,
-  onSuccess,
-}: {
-  isOpen: boolean;
-  onClose: () => void;
-  schema: TableSchema;
-  connection: Connection;
-  tableName: string;
-  onSuccess: () => void;
-}) => {
-  const [newRowData, setNewRowData] = useState<Record<string, any>>({});
-  const [isSaving, setIsSaving] = useState(false);
-
-  useEffect(() => {
-    if (isOpen) {
-      const initialData: Record<string, any> = {};
-      schema.columns.forEach((col) => {
-        if (col.defaultValue !== null && col.defaultValue !== undefined) {
-          initialData[col.columnName] = col.defaultValue;
-        } else if (col.isNullable) {
-          initialData[col.columnName] = null;
-        } else {
-          // Handle various data types with appropriate defaults
-          switch (col.dataType?.toLowerCase()) {
-            case "int":
-            case "integer":
-            case "bigint":
-            case "float":
-            case "double":
-            case "decimal":
-              initialData[col.columnName] = 0;
-              break;
-            case "boolean":
-            case "bool":
-              initialData[col.columnName] = false;
-              break;
-            default:
-              initialData[col.columnName] = "";
-          }
-        }
-      });
-      setNewRowData(initialData);
-    }
-  }, [isOpen, schema]);
-
-  const handleFieldChange = (columnName: string, value: any) => {
-    setNewRowData((prev) => ({ ...prev, [columnName]: value }));
-  };
-
-  const handleSubmit = async () => {
-    setIsSaving(true);
-    try {
-      const result = await insertRow(connection, tableName, newRowData);
-      if (result.success) {
-        toast.success("Row Added", { description: result.message });
-        onSuccess();
-        onClose();
-      } else {
-        toast.error("Add Row Failed", { description: result.message });
-      }
-    } catch (err) {
-      toast.error("Add Row Error", {
-        description: `An error occurred: ${(err as Error).message}`,
-      });
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  return (
-    <Dialog open={isOpen} onOpenChange={onClose}>
-      <DialogContent className="sm:max-w-[600px]">
-        <DialogHeader>
-          <DialogTitle>Add New Row to &quot;{tableName}&quot;</DialogTitle>
-          <DialogDescription>
-            Fill in the details for the new row.
-          </DialogDescription>
-        </DialogHeader>
-        <div className="grid gap-4 py-4 max-h-[60vh] overflow-y-auto pr-4">
-          {schema.columns.map((col) => (
-            <div
-              key={col.columnName}
-              className="grid grid-cols-4 items-center gap-4"
-            >
-              <Label htmlFor={col.columnName} className="text-right">
-                {col.columnName}
-                {!col.isNullable && <span className="text-destructive">*</span>}
-              </Label>
-              <div className="col-span-3">
-                <Input
-                  id={col.columnName}
-                  value={newRowData[col.columnName] ?? ""}
-                  onChange={(e) =>
-                    handleFieldChange(col.columnName, e.target.value)
-                  }
-                  placeholder={col.dataType}
-                  disabled={isSaving}
-                />
-              </div>
-            </div>
-          ))}
-        </div>
-        <DialogFooter>
-          <Button variant="outline" onClick={onClose} disabled={isSaving}>
-            Cancel
-          </Button>
-          <Button onClick={handleSubmit} disabled={isSaving}>
-            {isSaving ? "Saving..." : "Save Row"}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-};
 
 export function TableViewer() {
   const searchParams = useSearchParams();
